--- app/src/main/java/com/djacoronel/basiclauncher/CountUpTimer.java
@@ -1,45 +1,0 @@
-package com.djacoronel.basiclauncher;
-
-import android.os.Handler;
-import android.os.Message;
-import android.os.SystemClock;
-
-public abstract class CountUpTimer {
-
-    private final long interval;
-    private long base;
-
-    public CountUpTimer(long interval) {
-        this.interval = interval;
-    }
-
-    public void start() {
-        base = SystemClock.elapsedRealtime();
-        handler.sendMessage(handler.obtainMessage(MSG));
-    }
-
-    public void stop() {
-        handler.removeMessages(MSG);
-    }
-
-    public void reset() {
-        synchronized (this) {
-            base = SystemClock.elapsedRealtime();
-        }
-    }
-
-    abstract public void onTick(long elapsedTime);
-
-    private static final int MSG = 1;
-
-    private Handler handler = new Handler() {
-        @Override
-        public void handleMessage(Message msg) {
-            synchronized (CountUpTimer.this) {
-                long elapsedTime = SystemClock.elapsedRealtime() - base;
-                onTick(elapsedTime);
-                sendMessageDelayed(obtainMessage(MSG), interval);
-            }
-        }
-    };
-}
