+++ app/src/main/java/com/djacoronel/basiclauncher/DbHelper.java
@@ -25,33 +25,6 @@ class DbHelper extends SQLiteOpenHelper {
     private static final String SQL_DELETE_ENTRIES =
             "DROP TABLE IF EXISTS " + TABLE_NAME;
 
-    private static final String TABLE_NAME_TASKS = "tasks";
-    private static final String COLUMN_NAME_TNAME = "name";
-    private static final String COLUMN_NAME_TDURATION = "duration";
-    private static final String COLUMN_NAME_TSTATUS = "status";
-    private static final String COLUMN_NAME_TTYPE = "type";
-    private static final String COLUMN_NAME_TREMAINING = "remaining";
-
-    private static final String SQL_CREATE_TASKS =
-            "CREATE TABLE " + TABLE_NAME_TASKS + " (" +
-                    ID + " INTEGER PRIMARY KEY autoincrement," +
-                    COLUMN_NAME_TNAME + " TEXT," +
-                    COLUMN_NAME_TTYPE + " TEXT," +
-                    COLUMN_NAME_TDURATION + " TEXT," +
-                    COLUMN_NAME_TREMAINING + " TEXT," +
-                    COLUMN_NAME_TSTATUS + " TEXT)";
-
-    private static final String SQL_DELETE_TASKS =
-            "DROP TABLE IF EXISTS " + TABLE_NAME;
-
-    private static final String TABLE_NAME_WIDGETS = "widgets";
-    private static final String COLUMN_NAME_WID = "name";
-
-    private static final String SQL_CREATE_WIDGETS =
-            "CREATE TABLE " + TABLE_NAME_WIDGETS + " (" +
-                    ID + " INTEGER PRIMARY KEY AUTOINCREMENT," +
-                    COLUMN_NAME_WID + " TEXT)";
-
 
     DbHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -60,81 +33,17 @@ class DbHelper extends SQLiteOpenHelper {
     @Override
     public void onCreate(SQLiteDatabase db) {
         db.execSQL(SQL_CREATE_ENTRIES);
-        db.execSQL(SQL_CREATE_TASKS);
-        db.execSQL(SQL_CREATE_WIDGETS);
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
         db.execSQL(SQL_DELETE_ENTRIES);
-        db.execSQL(SQL_DELETE_TASKS);
-        db.execSQL(SQL_CREATE_WIDGETS);
         onCreate(db);
     }
 
     @Override
     public void onDowngrade(SQLiteDatabase db, int oldVersion, int newVersion) {
         onUpgrade(db, oldVersion, newVersion);
-    }
-
-    ArrayList<Task> getTasks(Context context) {
-        SQLiteDatabase db = getWritableDatabase();
-        Cursor cursor = db.rawQuery("select * from " + TABLE_NAME_TASKS, null);
-
-        ArrayList<Task> tasks = new ArrayList<>();
-        if (cursor.moveToFirst()) {
-            while (!cursor.isAfterLast()) {
-                Task task = new Task(cursor.getString(cursor.getColumnIndex(COLUMN_NAME_TNAME)),
-                        cursor.getString(cursor.getColumnIndex(COLUMN_NAME_TTYPE)),
-                        cursor.getString(cursor.getColumnIndex(COLUMN_NAME_TDURATION)),
-                        cursor.getString(cursor.getColumnIndex(COLUMN_NAME_TREMAINING)),
-                        cursor.getString(cursor.getColumnIndex(COLUMN_NAME_TSTATUS)),
-                        context);
-                task.setId(cursor.getLong(cursor.getColumnIndex(ID)));
-                tasks.add(task);
-                cursor.moveToNext();
-            }
-        }
-        cursor.close();
-        return tasks;
-    }
-
-    long addTask(Task task){
-        SQLiteDatabase db = getWritableDatabase();
-        ContentValues values = new ContentValues();
-        values.put(COLUMN_NAME_TNAME, task.getName());
-        values.put(COLUMN_NAME_TTYPE, task.getItemType());
-        values.put(COLUMN_NAME_TDURATION, task.getDuration());
-        values.put(COLUMN_NAME_TREMAINING, task.getTimeRemaining());
-        values.put(COLUMN_NAME_TSTATUS, task.getStatus());
-        return db.insert(TABLE_NAME_TASKS, null, values);
-    }
-
-    void updateTask(Task task){
-        SQLiteDatabase db = getWritableDatabase();
-
-        ContentValues values = new ContentValues();
-        values.put(COLUMN_NAME_TNAME, task.getName());
-        values.put(COLUMN_NAME_TTYPE, task.getItemType());
-        values.put(COLUMN_NAME_TDURATION, task.getDuration());
-        values.put(COLUMN_NAME_TREMAINING, task.getTimeRemaining());
-        values.put(COLUMN_NAME_TSTATUS, task.getStatus());
-
-        String selection = ID+ " LIKE ?";
-        String[] selectionArgs = { "" + task.getId() };
-
-        db.update(
-                TABLE_NAME_TASKS,
-                values,
-                selection,
-                selectionArgs);
-    }
-
-    void deleteTask(long id){
-        SQLiteDatabase db = getWritableDatabase();
-        String selection = ID + " LIKE ?";
-        String[] selectionArgs = {""+id};
-        db.delete(TABLE_NAME_TASKS, selection, selectionArgs);
     }
 
     ArrayList<String> getHiddenList() {
@@ -160,42 +69,5 @@ class DbHelper extends SQLiteOpenHelper {
         values.put(COLUMN_NAME_LABEL, label);
 
         return db.insert(TABLE_NAME, null, values);
-    }
-
-    void removeFromHidden(String label) {
-        SQLiteDatabase db = getWritableDatabase();
-        String selection = COLUMN_NAME_LABEL + " LIKE ?";
-        String[] selectionArgs = {label};
-        db.delete(TABLE_NAME, selection, selectionArgs);
-    }
-
-    ArrayList<String> getWidgets() {
-        SQLiteDatabase db = getWritableDatabase();
-        Cursor cursor = db.rawQuery("select * from " + TABLE_NAME_WIDGETS, null);
-
-        ArrayList<String> widgets = new ArrayList<>();
-        if (cursor.moveToFirst()) {
-            while (!cursor.isAfterLast()) {
-                widgets.add(cursor.getString(cursor.getColumnIndex(COLUMN_NAME_WID)));
-                cursor.moveToNext();
-            }
-        }
-        cursor.close();
-        return widgets;
-    }
-
-    long addWidget(int id) {
-        SQLiteDatabase db = getWritableDatabase();
-        ContentValues values = new ContentValues();
-        values.put(COLUMN_NAME_WID, "" + id);
-
-        return db.insert(TABLE_NAME_WIDGETS, null, values);
-    }
-
-    void removeWidget(int id) {
-        SQLiteDatabase db = getWritableDatabase();
-        String selection = COLUMN_NAME_WID + " LIKE ?";
-        String[] selectionArgs = {"" + id};
-        db.delete(TABLE_NAME_WIDGETS, selection, selectionArgs);
     }
 }
