--- app/src/main/java/com/djacoronel/basiclauncher/PagerAdapter.java
@@ -1,31 +1,0 @@
-package com.djacoronel.basiclauncher;
-
-
-import android.app.Fragment;
-import android.app.FragmentManager;
-import android.support.v13.app.FragmentPagerAdapter;
-
-public class PagerAdapter extends FragmentPagerAdapter {
-    int mNumOfTabs;
-
-    public PagerAdapter(FragmentManager fm, int NumOfTabs) {
-        super(fm);
-        this.mNumOfTabs = NumOfTabs;
-    }
-    @Override
-    public Fragment getItem(int position) {
-        switch (position) {
-            case 0:
-                return new TabFragment1();
-            case 1:
-                return new TabFragment2();
-            case 2:
-                return new TabFragment3();
-            default:
-                return null;
-        } }
-    @Override
-    public int getCount() {
-        return mNumOfTabs;
-    }
-}
