+++ app/src/main/res/layout/activity_main.xml
@@ -7,9 +7,17 @@
     android:layout_height="match_parent"
     tools:context="com.djacoronel.basiclauncher.MainActivity">
 
-    <android.support.v4.view.ViewPager
-        android:id="@+id/pager"
+    <include layout="@layout/app_grid" />
+
+    <View
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        />
+        android:layout_height="50dp"
+        android:layout_gravity="bottom"
+        android:background="@drawable/bottom_gradient" />
+
+    <View
+        android:layout_width="match_parent"
+        android:layout_height="50dp"
+        android:layout_gravity="top"
+        android:background="@drawable/top_gradient" />
 </FrameLayout>
