--- app/src/main/res/layout/tab_fragment1.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:id="@+id/coodinator">
-
-
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/task_list"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:clipToPadding="false"
-        android:paddingTop="50dp"
-        android:paddingBottom="50dp"
-        android:paddingStart="20dp"
-        android:paddingEnd="20dp"/>
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:layout_gravity="bottom"
-        android:background="@drawable/bottom_gradient" />
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:layout_gravity="top"
-        android:background="@drawable/top_gradient"/>
-
-</android.support.design.widget.CoordinatorLayout>
