--- app/src/main/res/layout/tab_fragment2.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:background="@android:color/background_dark"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.djacoronel.basiclauncher.MainActivity">
-
-    <include
-        layout="@layout/app_grid"/>
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:layout_gravity="bottom"
-        android:background="@drawable/bottom_gradient" />
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:layout_gravity="top"
-        android:background="@drawable/top_gradient"/>
-</FrameLayout>
