--- app/src/main/res/layout/tab_fragment3.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:id="@+id/widget_space"
-    android:paddingTop="30dp"
-    android:paddingBottom="30dp">
-
-</LinearLayout>
